+++ modules/extensions-xml-support/src/main/java/org/mule/runtime/extension/internal/loader/xml/validator/ForbiddenConfigurationPropertiesValidator.java
