+++ modules/extensions-xml-support/src/test/java/org/mule/runtime/extension/internal/loader/xml/validation/ConnectivityTestingFailuresTestCase.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.extension.internal.feeder.validation;
+package org.mule.runtime.extension.internal.loader.validation;
 
 
 import static java.lang.String.format;
@@ -15,9 +15,9 @@ import static org.hamcrest.Matchers.is;
 import static org.mule.runtime.api.dsl.DslResolvingContext.getDefault;
 import static org.mule.runtime.core.api.extension.provider.MuleExtensionModelProvider.getExtensionModel;
 import static org.mule.runtime.extension.api.loader.ExtensionModelLoadingRequest.builder;
-import static org.mule.runtime.extension.internal.feeder.XmlExtensionLoaderDelegate.MODULE_CONNECTION_MARKER_ANNOTATION_ATTRIBUTE;
-import static org.mule.runtime.extension.internal.feeder.XmlExtensionModelLoader.RESOURCE_XML;
-import static org.mule.runtime.extension.internal.feeder.validator.TestConnectionValidator.TEST_CONNECTION_SELECTED_ELEMENT_INVALID;
+import static org.mule.runtime.extension.internal.loader.xml.XmlExtensionLoaderDelegate.MODULE_CONNECTION_MARKER_ANNOTATION_ATTRIBUTE;
+import static org.mule.runtime.extension.internal.loader.xml.XmlExtensionModelLoader.RESOURCE_XML;
+import static org.mule.runtime.extension.internal.loader.xml.validator.TestConnectionValidator.TEST_CONNECTION_SELECTED_ELEMENT_INVALID;
 
 import org.mule.metadata.api.ClassTypeLoader;
 import org.mule.runtime.api.exception.MuleRuntimeException;
@@ -30,8 +30,8 @@ import org.mule.runtime.extension.api.declaration.type.ExtensionsTypeLoaderFacto
 import org.mule.runtime.extension.api.loader.ProblemsReporter;
 import org.mule.runtime.extension.internal.loader.DefaultExtensionLoadingContext;
 import org.mule.runtime.extension.internal.loader.ExtensionModelFactory;
-import org.mule.runtime.extension.internal.feeder.XmlExtensionModelLoader;
-import org.mule.runtime.extension.internal.feeder.validator.TestConnectionValidator;
+import org.mule.runtime.extension.internal.loader.xml.XmlExtensionModelLoader;
+import org.mule.runtime.extension.internal.loader.xml.validator.TestConnectionValidator;
 import org.mule.runtime.internal.dsl.NullDslResolvingContext;
 import org.mule.tck.junit4.AbstractMuleTestCase;
 import org.mule.tck.size.SmallTest;
